backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
apscheduler>=3.10.4
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-dateutil>=2.8.2
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4